Fix constant propagation to not delete the exit block of a function

+++ src/main/java/com/d_m/pass/ConstantPropagation.java
@@ -50,7 +50,7 @@ public class ConstantPropagation extends BooleanFunctionPass {
                 continue;
             }
             changes |= removeSingleArgumentPhis(block);
-            if (isEmptyBlock(block)) {
+            if (isEmptyBlock(block) && !block.equals(block.getExit())) {
                 changes = true;
                 deleteEmptyBlock(block);
                 blocksToRemove.add(block);
